Improve class organization and some others points
