+++ belvo/resources/__init__.py
@@ -1,0 +1,12 @@
+from belvo.resources.accounts import Accounts
+from belvo.resources.balances import Balances
+from belvo.resources.incomes import Incomes
+from belvo.resources.institutions import Institutions
+from belvo.resources.invoices import Invoices
+from belvo.resources.links import Links
+from belvo.resources.owners import Owners
+from belvo.resources.statements import Statements
+from belvo.resources.tax_returns import TaxReturns
+from belvo.resources.tax_status import TaxStatus
+from belvo.resources.transactions import Transactions
+from belvo.resources.widget_token import WidgetToken
