+++ belvo/resources/incomes.py
@@ -1,0 +1,43 @@
+from typing import Union, List, Dict, Generator
+
+from belvo.base_resource import Resource
+
+
+class Incomes(Resource):
+    endpoint = "/api/incomes/"
+
+    def create(
+        self,
+        link: str,
+        *,
+        token: str = None,
+        encryption_key: str = None,
+        save_data: bool = True,
+        raise_exception: bool = False,
+        **kwargs: Dict,
+    ) -> Union[List[Dict], Dict]:
+
+        data = {"link": link, "save_data": save_data}
+
+        if token:
+            data.update(token=token)
+        if encryption_key:
+            data.update(encryption_key=encryption_key)
+
+        return self.session.post(
+            self.endpoint, data=data, raise_exception=raise_exception, **kwargs
+        )
+
+    def list(self, **kwargs) -> Generator:
+        raise NotImplementedError()
+
+    def get(self, id: str, **kwargs) -> Dict:
+        raise NotImplementedError()
+
+    def delete(self, id: str) -> bool:
+        raise NotImplementedError()
+
+    def resume(
+        self, session: str, token: str, *, link: str = None, raise_exception: bool = False, **kwargs
+    ) -> Union[List[Dict], Dict]:
+        raise NotImplementedError()
